requirements.txt
flask>=3.0.0
aws-wsgi==0.2.7
python-dotenv>=1.0.0
boto3>=1.34.0
pylint>=3.0.0
pytest>=7.0.0
